BREAKING: change how we deal with environment configuration
The environment configuration is now shared, and includes information about OIDC
config and endpoints.

Added cli command for configuring the environment that can be used by consuming
applications.

Tokens are still set and stored on a per application basis.

Now using a filesystem lock to prevent simultaneous operations on the
configuration files.